+++ Packages/Sources/TileKit/TileKit.swift
@@ -1,6 +1,7 @@
 @_exported import TileContent
 @_exported import TileCore
 @_exported import TileMarkdown
+@_exported import TileService
 @_exported import TileSite
 @_exported import TileSiteImpl
 @_exported import TileSource